rename: Spring 설정 클래스 패키지 이동
