+++ src/main/java/com/tierlist/tierlist/global/security/config/PasswordEncoderConfig.java
@@ -1,4 +1,4 @@
-package com.tierlist.tierlist.global.security;
+package com.tierlist.tierlist.global.security.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
